+++ server/src/main/java/project/manager/server/dto/exception/ExceptionDto.java
@@ -1,4 +1,4 @@
-package project.manager.server.dto;
+package project.manager.server.dto.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
